+++ desktop_app/src/backend/server/plugins/oauth/provider-interface.ts
