+++ desktop_app/src/backend/server/plugins/oauth/utils/oauth-config.ts
